Example of Unit Test: Isolation, dependent test (using stub and Mock), integration test
Example of Unit Test: Isolation, dependent test (using stub and Mock),
integration test

+++ spring-testing/src/main/java/com/dee/spring/testing/basic1/impl/StudentServiceImpl.java
@@ -8,17 +8,17 @@ import com.dee.spring.testing.basic1.dao.StudentDao;
  * @author dien.nguyen
  **/
 
-public class InMemoryStudentService implements StudentService{
+public class StudentServiceImpl implements StudentService{
     
     private StudentDao studentDao;
     
-    public InMemoryStudentService(StudentDao studentDao) {
+    public StudentServiceImpl(StudentDao studentDao) {
         this.studentDao = studentDao;
     }
 
     @Override
     public Student getById(Long id) {
-        return null;
+        return studentDao.findById(id);
     }
 
     @Override
@@ -32,8 +32,20 @@ public class InMemoryStudentService implements StudentService{
     }
 
     @Override
-    public void remove(Student student) {
-        studentDao.delete(student);
+    public void remove(Long id) {
+        Student student = studentDao.findById(id);
+        if(student != null) {
+            studentDao.delete(student);
+        }
+    }
+
+    @Override
+    public Student updateActive(Long id, boolean active) {
+        Student student = studentDao.findById(id);
+        student.setActive(active);
+        studentDao.update(student);
+        
+        return student;
     }
 
 }
